refactor: 에러 이벤트 위치 수정 & 쿠폰 이벤트 추가
