+++ app/src/main/java/woowacourse/shopping/data/repository/ProductRepositoryImpl.kt
@@ -4,7 +4,7 @@ import woowacourse.shopping.data.db.product.MockProductService
 import woowacourse.shopping.domain.model.Product
 import woowacourse.shopping.domain.repository.ProductRepository
 import woowacourse.shopping.domain.service.ProductService
-import woowacourse.shopping.view.model.event.ErrorEvent
+import woowacourse.shopping.utils.exception.ErrorEvent
 import kotlin.concurrent.thread
 
 class ProductRepositoryImpl : ProductRepository {
