+++ app/src/main/java/woowacourse/shopping/data/repository/RecentlyProductRepositoryImpl.kt
@@ -6,7 +6,7 @@ import woowacourse.shopping.domain.model.RecentlyProduct
 import woowacourse.shopping.domain.repository.RecentlyProductRepository
 import woowacourse.shopping.utils.EntityMapper.toRecentlyProduct
 import woowacourse.shopping.utils.EntityMapper.toRecentlyProductEntity
-import woowacourse.shopping.view.model.event.ErrorEvent
+import woowacourse.shopping.utils.exception.ErrorEvent
 import kotlin.concurrent.thread
 
 class RecentlyProductRepositoryImpl(context: Context) : RecentlyProductRepository {
