+++ app/src/main/java/woowacourse/shopping/view/BaseViewModel.kt
@@ -3,7 +3,7 @@ package woowacourse.shopping.view
 import androidx.lifecycle.ViewModel
 import woowacourse.shopping.utils.livedata.MutableSingleLiveData
 import woowacourse.shopping.utils.livedata.SingleLiveData
-import woowacourse.shopping.view.model.event.ErrorEvent
+import woowacourse.shopping.utils.exception.ErrorEvent
 
 open class BaseViewModel : ViewModel() {
     private val _errorEvent: MutableSingleLiveData<ErrorEvent> = MutableSingleLiveData()
