+++ app/src/main/java/woowacourse/shopping/view/recommend/RecommendViewModel.kt
@@ -21,7 +21,7 @@ import woowacourse.shopping.utils.livedata.SingleLiveData
 import woowacourse.shopping.view.BaseViewModel
 import woowacourse.shopping.view.cart.model.ShoppingCart
 import woowacourse.shopping.view.cartcounter.OnClickCartItemCounter
-import woowacourse.shopping.view.model.event.ErrorEvent
+import woowacourse.shopping.utils.exception.ErrorEvent
 
 class RecommendViewModel(
     private val orderRepository: OrderRepository,
